test: shared cross-language conformance fixtures + real-catalog guard
- host-protocol/fixtures/inbound-messages.json: one source of truth both the TS
  (messages.spec) and Kotlin (InboundConformanceTest) parsers assert against, so
  the two parsers can't silently drift (the audit's #1).
- The docVersion>2^31 fixture surfaced a real bug: the Kotlin parser truncated
  docVersion via getInt. Fixed to Long end-to-end (HostProtocol/dispatcher/editor).
- CommandSurfaceTest now also asserts parseCommandMenu over the REAL commands.json,
  guarding the kebab against catalog drift (the audit's #2).
- :protocol Gradle copies the shared fixtures + commands.json into test resources.

+++ arrows-code/hosts/intellij/plugin/src/main/kotlin/app/arrows/intellij/ArrowsFileEditor.kt
@@ -174,7 +174,7 @@ class ArrowsFileEditor(
 
     override fun onReady() = sendLoad()
 
-    override fun onGraphChanged(graph: GraphPayload, docVersion: Int?) {
+    override fun onGraphChanged(graph: GraphPayload, docVersion: Long?) {
         val doc = document ?: return
         // Write the graph back verbatim - preserves style and every top-level
         // field. (Canonical key ordering would need the bundle to emit it.)
